Add permission for managing navigation links
Users with manage_nav_menu_links permission can
add and remove links. Others can still rearrange.

closes INTEROP-10081
flag=nav_menu_links

test plan:
- Run the datafixup
- as account admin, go to account settings
- verify can add/remove navigation links
- Add permission to Teacher role
- as teacher with permission, go to course settings
- verify can add/remove course navigation links
- remove permission from teacher role
- verify cannot add/remove but can rearrange links
- verify root admin always has permission
- Create custom Admin Role without this permission
- Make sure Admin can see but cannot update links

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/401846
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Migration-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>

+++ lib/nav_menu_link_tabs.rb
@@ -64,7 +64,8 @@ module NavMenuLinkTabs
   # * Creates new links (with linkUrl and no id) and replaces them with {id: "nav_menu_link_123"}
   # * Filtering out any links that are for the wrong context / nav type
   # (Note that course-context course all have only course_nav: true)
-  def sync_course_links_with_tabs(course:, tabs:)
+  # @param can_manage_links [Boolean] If false, skip creating/deleting links (but allow rearranging)
+  def sync_course_links_with_tabs(course:, tabs:, can_manage_links: false)
     raise ArgumentError unless course.is_a?(Course)
 
     tabs = tabs.map(&:with_indifferent_access)
@@ -86,10 +87,16 @@ module NavMenuLinkTabs
         new_link_url = (tab[:args].is_a?(Array) && tab[:args][0].is_a?(String)) ? tab[:args][0] : nil
 
         if id.nil? && tab[:href].to_s == TAB_HREF_VALUE.to_s && new_link_url.present?
-          new_link = NavMenuLink.create!(context: course, course_nav: true, url: new_link_url, label: tab[:label])
-          { id: numeric_id_to_tab_json_id(new_link.id), hidden: tab[:hidden] }.compact.with_indifferent_access
+          if can_manage_links
+            new_link = NavMenuLink.create!(context: course, course_nav: true, url: new_link_url, label: tab[:label])
+            { id: numeric_id_to_tab_json_id(new_link.id), hidden: tab[:hidden] }.compact.with_indifferent_access
+          else
+            # User lacks permission to create links, skip this tab
+            Rails.logger.info("NavMenuLinkTabs.sync_course_links_with_tabs: Skipping link creation due to lack of permission")
+            nil
+          end
         elsif id.nil?
-          Rails.logger.warn("NavMenuLinkTabs.sync_course_links_with_tabs: Ignoring invalid tab: #{tab.inspect}")
+          Rails.logger.warn("NavMenuLinkTabs.sync_course_links_with_tabs: Ignoring invalid tab: (id is nil)")
           nil
         elsif numeric_id.nil? || current_valid_link_ids.include?(numeric_id)
           # Non-link, or reference to valid link in our account chain
@@ -100,7 +107,12 @@ module NavMenuLinkTabs
         end
       end
 
-      NavMenuLink.where(id: course_links_to_delete).destroy_all
+      # Only delete links if user has permission
+      if can_manage_links && course_links_to_delete.any?
+        NavMenuLink.where(id: course_links_to_delete).destroy_all
+      elsif !can_manage_links && course_links_to_delete.any?
+        Rails.logger.info("NavMenuLinkTabs.sync_course_links_with_tabs: Skipping link deletion due to lack of permission")
+      end
 
       result
     end
